third_party/WebKit/LayoutTests/inspector/elements/styles/dynamic-style-tag.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<style>
/* comment */.inline-style-added-by-parser {
   color: red;
}
</style>
<style>
.inline-style-added-by-parser-with-source-url {
   color: green;
}
/*# sourceURL=inlineStyleAddedByParser.css*/
</style>
<script>

function addStyleElements()
{
    addStyleElement(".inline-style-created-by-script {\n   color: orange;\n}");
    addStyleElement(".inline-style-created-by-script-with-source-url {\n   color: grey;\n}\n/*# sourceURL=inlineStyleCreatedByScript.css*/");
}

function addStyleElement(styleContent)
{
    var styleElement = document.createElement("style");
    styleElement.textContent = styleContent;
    document.head.appendChild(styleElement);
}

function test()
{
    WebInspector.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.StyleSheetAdded, styleSheetAdded);
    InspectorTest.evaluateInPage("addStyleElements()", step1);

    function step1()
    {
        InspectorTest.selectNodeAndWaitForStyles("inspected", step2);
    }

    function step2()
    {
        InspectorTest.dumpSelectedElementStyles(true, false, true);
        InspectorTest.completeTest();
    }

    function styleSheetAdded(event)
    {
        var header = event.data;
        var text = ["Stylesheet added:"];
        text.push("  - isInline: " + header.isInline);
        text.push("  - sourceURL: " + pathToFileName(header.sourceURL));
        text.push("  - hasSourceURL: " + header.hasSourceURL);

        CSSAgent.getStyleSheetText(header.id, contentLoaded.bind(this));

        function pathToFileName(path)
        {
            return path.substring(path.lastIndexOf("/") + 1);
        }

        function contentLoaded(error, content)
        {
            text.push("  - contents: " + content);
            InspectorTest.addResult(text.join("\n"));
        }
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that different types of inline styles are correctly disambiguated and their sourceURL is correct.
<div id="inspected" style="color:red" class="inline-style-added-by-parser inline-style-added-by-parser-with-source-url inline-style-added-by-parser-in-document-write inline-style-added-by-document-write-with-source-url inline-style-created-by-script inline-style-created-by-script-with-source-url"></div>
</body>
</html>
