third_party/WebKit/LayoutTests/inspector/elements/styles/dynamic-style-tag-expected.txt

Tests that different types of inline styles are correctly disambiguated and their sourceURL is correct.

Stylesheet added:
  - isInline: false
  - sourceURL: 
  - hasSourceURL: false
  - contents: .inline-style-created-by-script {
   color: orange;
}
Stylesheet added:
  - isInline: false
  - sourceURL: inlineStyleCreatedByScript.css
  - hasSourceURL: true
  - contents: .inline-style-created-by-script-with-source-url {
   color: grey;
}
/*# sourceURL=inlineStyleCreatedByScript.css*/
[expanded] 
element.style  { ()
color: red;

[expanded] 
.inline-style-created-by-script-with-source-url  { (inlineStyleCrea…ByScript.css:1 -> inlineStyleCreatedByScript.css:1:1)
/-- overloaded --/ color: grey;

[expanded] 
.inline-style-created-by-script  { ()
/-- overloaded --/ color: orange;

[expanded] 
.inline-style-added-by-parser-with-source-url  { ()
/-- overloaded --/ color: green;

[expanded] 
.inline-style-added-by-parser  { (dynamic-style-tag.html:6 -> dynamic-style-tag.html:6:14)
/-- overloaded --/ color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;
